Db update

+++ src/program/UserPanelController.java
@@ -1,5 +1,6 @@
 package program;
 
+import java.sql.Connection;
 import java.sql.ResultSet;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -45,18 +46,22 @@ public class UserPanelController {
     Alert alert2 = new Alert(AlertType.ERROR);
     alert2.setTitle("Error!");
     alert2.setHeaderText(null);
+    ResultSet rs = null;
+    Connection conn = null;
 
     String query = "SELECT * FROM ACCOUNT WHERE USERNAME='" + username.getText() + "'";
     try {
+      conn = DBconn.derbyConn();
+
       //run query and save result set
-      ResultSet rs = DBconn.queryDb(query, DBconn.derbyConn());
+      rs = DBconn.queryDb(query, conn);
       if (rs == null || !rs.next()) {
         //if the result set has at least one value then it means the value inserted was valid
         //then update with new value typed by the user
         //create query
         query = "INSERT INTO ACCOUNT(name, username, password) VALUES ('" + name.getText()
             + "', '" + username.getText() + "', '" + password.getText() + "')";
-        if (DBconn.updateDb(query, DBconn.derbyConn()) != -1) {
+        if (DBconn.updateDb(query, conn) != -1) {
           alert.setContentText("New account created successfully!");
           alert.showAndWait();
 
@@ -78,6 +83,13 @@ public class UserPanelController {
       }
     } catch (Exception e) {
       System.out.println(e.toString());
+    } finally {
+      try {
+        rs.close();
+        conn.close();
+      } catch (Exception e) {
+        e.printStackTrace();
+      }
     }
   }
 
@@ -95,43 +107,62 @@ public class UserPanelController {
     Alert alert2 = new Alert(AlertType.ERROR);
     alert2.setTitle("Error!");
     alert2.setHeaderText(null);
+    ResultSet rs = null;
+    Connection conn = null;
 
     //if deletioncheck is selected it means the user is sure to remove this account
     if (deletionCheck.isSelected()) {
+      if (!usernameDel.getText().equals(Main.loggedUser)) {
+        //check that the user exists
+        String query = "SELECT * FROM ACCOUNT WHERE USERNAME='" + usernameDel.getText() + "'";
+        try {
+          conn = DBconn.derbyConn();
+          //run query and save result set
+          rs = DBconn.queryDb(query, conn);
+          if (rs != null) {
+            //if the result set has at least one value then it means the value inserted was valid
+            if (rs.next()) {
+              //create query to delete account
+              query = "DELETE FROM ACCOUNT WHERE USERNAME='" + usernameDel.getText() + "'";
+              //run query
+              if (DBconn.updateDb(query, conn) != -1) {
+                alert.setContentText("Account deleted successfully!");
+                alert.showAndWait();
 
-      //check that the user exists
-      String query = "SELECT * FROM ACCOUNT WHERE USERNAME='" + usernameDel.getText() + "'";
-      try {
-        //run query and save result set
-        ResultSet rs = DBconn.queryDb(query, DBconn.derbyConn());
-        if (rs != null) {
-          //if the result set has at least one value then it means the value inserted was valid
-          if (rs.next()) {
-            //create query to delete account
-            query = "DELETE FROM ACCOUNT WHERE USERNAME='" + usernameDel.getText() + "'";
-            //run query
-            if (DBconn.updateDb(query, DBconn.derbyConn()) != -1) {
-              alert.setContentText("Account deleted successfully!");
-              alert.showAndWait();
+                //reset textField and deselect deletionCheck for next use
+                usernameDel.setText("");
+                deletionCheck.setSelected(false);
 
-              //reset textField and deselect deletionCheck for next use
-              usernameDel.setText("");
-              deletionCheck.setSelected(false);
-
-              //return to home screen
-              Main.setPane(Screens.HOME.getValue());
+                //return to home screen
+                Main.setPane(Screens.HOME.getValue());
+              } else {
+                System.err.println("FAILED TO RUN QUERY");
+              }
             } else {
-              System.err.println("FAILED TO RUN QUERY");
+              alert2.setContentText("Account not found, double check username!");
+              alert2.showAndWait();
+              usernameDel.clear();
             }
-          } else {
-            alert2.setContentText("Account not found, double check username!");
-            alert2.showAndWait();
-            usernameDel.clear();
+          }
+        } catch (Exception e) {
+          System.out.println(e.toString());
+        } finally {
+          try {
+            rs.close();
+            conn.close();
+          } catch (Exception e) {
+            e.printStackTrace();
           }
         }
-      } catch (Exception e) {
-        System.out.println(e.toString());
+      } else {
+        alert2.setContentText("You can't delete yourself, dork!");
+        alert2.showAndWait();
+        usernameDel.clear();
       }
+    } else {
+      alert2.setContentText("Tick the checkbox to ensure that you want to delete the user!");
+      alert2.showAndWait();
+      usernameDel.clear();
     }
 
   }
